chore: fix test coverage
